requirements.txt
agno==1.5.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.9
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-genai==1.18.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
groq==0.26.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.31
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.74.0
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
posthog==3.24.1
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic-bin
python-multipart==0.0.20
PyQt6==6.9.0
PyQt6-Qt6==6.9.1
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
PyQt6_sip==13.10.2
PySide6>=6.5
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
shellingham==1.5.4
shiboken6==6.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
unstructured[pdf]
codecarbon